foto

+++ model/db.php
@@ -18,9 +18,19 @@ class Banco{
     }
 
     public function setFuncionario($nome,$cpf,$email,$telefone,$endereco,$cidade){
-        $stmt = $this->mysqli->prepare("INSERT INTO funcionarios ( `nome`, `cpf`, `email`, `telefone`, `endereco`,`cidade`) VALUES (?,?,?,?,?,?)");
-        $stmt->bind_param("ssssss",$nome,$cpf,$email,$telefone,$endereco,$cidade);
-         if( $stmt->execute() == TRUE){
+
+        if (isset($_FILES['foto'])){
+            $nome_foto = $_FILES['foto']['name'];
+            $tmp_name = $_FILES['foto']['tmp_name'];
+            $location = '../arquivos/'. $_FILES['foto']['name'];;
+        
+            move_uploaded_file($tmp_name,$location);
+        }
+        
+        $stmt = $this->mysqli->prepare("INSERT INTO funcionarios ( `foto`,`nome`, `cpf`, `email`, `telefone`, `endereco`,`cidade`) VALUES (?,?,?,?,?,?,?)");
+        $stmt->bind_param("sssssss",$nome_foto,$nome,$cpf,$email,$telefone,$endereco,$cidade);
+       
+        if( $stmt->execute() == TRUE){
             return true ;
         }else{
             return false;
@@ -41,11 +51,20 @@ class Banco{
     }
 
     public function deleteFuncionario($id){
-        if($this->mysqli->query("DELETE FROM `funcionarios` WHERE `id` = '".$id."';")== TRUE){
-            return true;
-        }else{
-            return false;
+        $location = "../arquivos/"; 
+        $result = $this->mysqli->query("SELECT foto FROM funcionarios WHERE id = $id");
+        $foto = mysqli_fetch_object($result);
+        $imagemQueVaiDeletada = $location . $foto->foto; 
+        $deleta = unlink($imagemQueVaiDeletada);
+        if($deleta){
+            $this->mysqli->query("DELETE FROM tabela WHERE id = $id"); 
+            if($this->mysqli->query("DELETE FROM `funcionarios` WHERE `id` = '".$id."';")== TRUE){
+                return true;
+            }else{
+                return false;
+            }
         }
+        
 
     }
     public function pesquisaFuncionario($id){
@@ -55,8 +74,15 @@ class Banco{
 
     }
      public function updateFuncionario($nome,$cpf,$email,$telefone,$endereco,$cidade,$id){
-        $stmt = $this->mysqli->prepare("UPDATE `funcionarios` SET `nome` = ?, `cpf`=?, `email`=?, `telefone`=?, `endereco`=?,`cidade` = ? WHERE `id` = ?");
-        $stmt->bind_param("sssssss",$nome,$cpf,$email,$telefone,$endereco,$cidade,$id);
+        if (isset($_FILES['foto'])){
+            $nome_foto = $_FILES['foto']['name'];
+            $tmp_name = $_FILES['foto']['tmp_name'];
+            $location = '../arquivos/'. $_FILES['foto']['name'];;
+        
+            move_uploaded_file($tmp_name,$location);
+        }
+        $stmt = $this->mysqli->prepare("UPDATE `funcionarios` SET `foto` = ?,`nome` = ?, `cpf`=?, `email`=?, `telefone`=?, `endereco`=?,`cidade` = ? WHERE `id` = ?");
+        $stmt->bind_param("ssssssss",$nome_foto,$nome,$cpf,$email,$telefone,$endereco,$cidade,$id);
         if($stmt->execute()==TRUE){
             return true;
         }else{
